Wrap the serial port
Introduce a new type `TimeoutSerialPort` that implements `Read` by
reading from the serial port with a decreasing timeout.

Next, wrap this is a `BufReader`. Together, this gives buffered reading
with a timeout which simplifies reading lines and packets.

This complicates tracing since now reading from the serial port happens
inside a struct without access to the observers. Fix that by introducing
an `EventProvider` which owns the observers and can be cloned. Both
clones will use the same observers.